fix: import USDC Storage section Framer Motion's components on client-side
